+++ lib/ts/recipe/thirdparty/components/themes/signInAndUpCallback/index.tsx
@@ -21,6 +21,7 @@ import { PureComponent } from "react";
 import SpinnerIcon from "../../../../../components/assets/spinnerIcon";
 import StyleContext from "../../../../../styles/styleContext";
 import { withOverride } from "../../../../../components/componentOverride/withOverride";
+import { SuperTokensBranding } from "../../../../../components/SuperTokensBranding";
 /*
  * Component.
  */
@@ -42,6 +43,7 @@ class ThirdPartySignInAndUpCallbackTheme extends PureComponent {
                         <SpinnerIcon color={styles.palette.colors.primary} />
                     </div>
                 </div>
+                <SuperTokensBranding />
             </div>
         );
     };